requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.0
flask>=2.0.0
argparse>=1.4.0
python-dotenv>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.15.0